refactor: split oversized files, raise max-file-lines thresholds to 300/500
File splits (4 files → 10):
- state/stores.ts → cursor-store, persistent-store, scoped-store
- tooling/tsdoc-analysis.ts → tsdoc-analysis + tsdoc-package
- mcp/server.ts → extracted handler-adapters.ts
- cli/input-parsers.ts → input-parsers + input-file

Threshold changes:
- max-file-lines warn: 200 → 300, error: 400 → 500
- Updated tests to match new default thresholds

Suppressions with inline justification (>500 LOC, no extractable seams):
- cli/types.ts (1030 LOC) — types-only file
- mcp/server.ts (784 LOC) — cohesive server factory
- cli/command.ts (546 LOC) — fluent builder pattern
- cli/presets.ts (508 LOC) — uniform preset collection

Also: removed stale eslint-disable max-file-lines comments from
colors.ts and logging/types.ts (now below threshold), fixed
oxlint-disable-next-line positioning in contracts/envelope.ts.

🤘🏻 In-collaboration-with: [Claude Code](https://claude.com/claude-code)
