+++ packages/cli/src/colors/colors.ts
@@ -1,4 +1,3 @@
-/* eslint-disable outfitter/max-file-lines -- Nested colors/ dir breaks bunup shared chunk paths; 411 LOC barely over threshold */
 /**
  * Color tokens and ANSI utilities.
  *
