+++ packages/cli/src/types.ts
@@ -1,4 +1,4 @@
-/* eslint-disable outfitter/max-file-lines -- Types-only file; splitting interfaces across files hurts discoverability */
+/* oxlint-disable outfitter/max-file-lines -- Types-only file (30+ interfaces with TSDoc); splitting hurts discoverability */
 /**
  * Type definitions for @outfitter/cli
  *
@@ -664,7 +664,7 @@ export interface PaginationPresetConfig {
 /**
  * Resolved interaction flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type InteractionFlags = {
   /** Whether interactive prompts are allowed */
   readonly interactive: boolean;
@@ -676,7 +676,7 @@ export type InteractionFlags = {
 /**
  * Resolved strict mode flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type StrictFlags = {
   /** Whether strict mode is enabled */
   readonly strict: boolean;
@@ -685,7 +685,7 @@ export type StrictFlags = {
 /**
  * Resolved time-window flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type TimeWindowFlags = {
   /** Start of time window */
   readonly since: Date | undefined;
@@ -705,7 +705,7 @@ export interface TimeWindowPresetConfig {
 /**
  * Resolved execution flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type ExecutionFlags = {
   /** Timeout in milliseconds (undefined = no timeout) */
   readonly timeout: number | undefined;
@@ -733,7 +733,7 @@ export interface ExecutionPresetConfig {
 /**
  * Resolved projection flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type ProjectionFlags = {
   /** Fields to include (undefined = all) */
   readonly fields: string[] | undefined;
@@ -753,7 +753,7 @@ export type ColorMode = "auto" | "always" | "never";
 /**
  * Resolved color flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type ColorFlags = {
   /** Color output mode */
   readonly color: ColorMode;
@@ -762,7 +762,7 @@ export type ColorFlags = {
 /**
  * Resolved pagination flags from CLI input.
  */
-// eslint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
+// oxlint-disable-next-line typescript/consistent-type-definitions -- must be `type` to satisfy Record<string, unknown> constraint in FlagPreset<T>
 export type PaginationFlags = {
   /** Number of results to return */
   readonly limit: number;
@@ -1013,7 +1013,7 @@ export interface CursorOptions {
  * Re-export error classes from contracts for convenience.
  * These are the canonical error types for the CLI.
  */
-// eslint-disable-next-line oxc/no-barrel-file -- intentional re-exports for API surface
+// oxlint-disable-next-line oxc/no-barrel-file -- intentional re-exports for API surface
 export {
   CancelledError,
   type ErrorCategory,
