+++ packages/contracts/src/envelope.ts
@@ -114,14 +114,12 @@ export function toEnvelope<T, E extends OutfitterError>(
     };
   }
 
+  // oxlint-disable-next-line outfitter/no-process-env-in-packages -- boundary: production detection for stack-trace stripping
+  const isProduction = process.env["NODE_ENV"] === "production";
+
   return {
     ok: false,
-    // eslint-disable-next-line outfitter/no-process-env-in-packages -- boundary: production detection for stack-trace stripping
-    error: serializeError(
-      result.error,
-      undefined,
-      process.env["NODE_ENV"] === "production"
-    ),
+    error: serializeError(result.error, undefined, isProduction),
     meta: envelopeMeta,
   };
 }
