+++ packages/logging/src/internal/types.ts
@@ -1,4 +1,3 @@
-/* eslint-disable outfitter/max-file-lines -- Types-only file (14+ interfaces with TSDoc); splitting hurts discoverability */
 import type { Logger as ContractLogger } from "@outfitter/contracts";
 
 // ============================================================================
